requirements.txt
albumentations==1.3.0
efficientnet_pytorch==0.7.1
imageio==2.9.0
jpegio==0.2.8
lmdb==1.2.1
matplotlib==3.3.4
mmcv==1.6.0
mmcv_full==1.7.0
mmsegmentation==0.28.0
numpy==1.20.1
Pillow==9.5.0
pytorch_lightning==1.7.7
scipy==1.9.1
segmentation_models_pytorch==0.2.1
six==1.15.0
timm==0.4.12
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
yacs==0.1.8